Compute a rough polygon outline

Compute a rough polygon outline

Usage

rough_polygon(x, y = NULL, hand = NULL)

Arguments

x, y

Polygon coordinates.

hand

Hand-drawn geometry settings created with hand().

Value

A list with x and y components containing a roughened closed outline suitable for plotting with graphics::lines().

Examples

rough_polygon(c(2, 5, 8, 3), c(2, 7, 5, 1))
#> $x
#>   [1] 2.000000 2.047270 2.094470 2.141593 2.188632 2.235579 2.282429 2.329175
#>   [9] 2.375810 2.422330 2.468729 2.515003 2.561146 2.607154 2.653025 2.698753
#>  [17] 2.744338 2.789776 2.835064 2.880202 2.925187 2.970019 3.014698 3.059889
#>  [25] 3.107246 3.154583 3.201881 3.249123 3.296290 3.343365 3.390329 3.437163
#>  [33] 3.483852 3.530376 3.576718 3.622862 3.668790 3.714488 3.759939 3.805127
#>  [41] 3.850040 3.894662 3.938982 3.982986 4.026663 4.070004 4.111166 4.149982
#>  [49] 4.188673 4.227253 4.265741 4.304152 4.342504 4.380811 4.419092 4.457361
#>  [57] 4.495634 4.533927 4.572255 4.610632 4.649072 4.687590 4.726199 4.764911
#>  [65] 4.803739 4.842695 4.881789 4.921031 4.960432 5.000000 5.073150 5.146336
#>  [73] 5.219537 5.292735 5.365912 5.439046 5.512120 5.585113 5.658006 5.730781
#>  [81] 5.803419 5.875901 5.948211 6.020331 6.091798 6.162906 6.233749 6.304326
#>  [89] 6.374635 6.444675 6.514447 6.583955 6.653203 6.722195 6.790940 6.859445
#>  [97] 6.927719 6.995773 7.063704 7.131480 7.199059 7.266455 7.333679 7.400747
#> [105] 7.467674 7.534476 7.601168 7.667768 7.734293 7.800761 7.867189 7.933596
#> [113] 8.000000 7.931367 7.862694 7.793988 7.725252 7.656492 7.587715 7.518924
#> [121] 7.450126 7.381327 7.312532 7.243747 7.174978 7.106232 7.037512 6.968827
#> [129] 6.900182 6.831582 6.763034 6.694544 6.626117 6.557760 6.489478 6.421276
#> [137] 6.353161 6.285138 6.218891 6.152968 6.087222 6.021648 5.956240 5.890993
#> [145] 5.825902 5.760959 5.696159 5.631494 5.566957 5.502542 5.438239 5.374042
#> [153] 5.309943 5.245933 5.182004 5.118148 5.054356 4.990620 4.926931 4.863281
#> [161] 4.799660 4.736061 4.672475 4.607075 4.541427 4.475887 4.410457 4.345141
#> [169] 4.279938 4.214852 4.149883 4.085032 4.020300 3.955687 3.891192 3.826817
#> [177] 3.762559 3.698418 3.634394 3.570483 3.506684 3.442996 3.379415 3.315938
#> [185] 3.252563 3.189285 3.126102 3.063008 3.000000 2.935508 2.871076 2.806780
#> [193] 2.742688 2.678866 2.615694 2.553060 2.490824 2.428941 2.367352 2.305682
#> [201] 2.243924 2.182508 2.121411 2.060594 2.000000
#> 
#> $y
#>   [1] 2.000000 2.070189 2.140420 2.210696 2.281024 2.351406 2.421847 2.492350
#>   [9] 2.562920 2.633558 2.704270 2.775056 2.845921 2.916867 2.987895 3.059009
#>  [17] 3.130209 3.201497 3.272875 3.344343 3.415902 3.487554 3.559297 3.630733
#>  [25] 3.700870 3.771018 3.841190 3.911396 3.981646 4.051952 4.122325 4.192774
#>  [33] 4.263312 4.333949 4.404694 4.475558 4.546552 4.617684 4.688964 4.760402
#>  [41] 4.832005 4.903782 4.975741 5.047890 5.120234 5.192780 5.266634 5.341895
#>  [49] 5.417231 5.492634 5.568092 5.643596 5.719136 5.794702 5.870284 5.945873
#>  [57] 6.021460 6.097035 6.172589 6.248114 6.323600 6.399040 6.474425 6.549749
#>  [65] 6.625003 6.700180 6.775275 6.850280 6.925190 7.000000 6.958563 6.917178
#>  [73] 6.875817 6.834452 6.793054 6.751593 6.710040 6.668367 6.626544 6.584544
#>  [81] 6.542337 6.499898 6.457200 6.414218 6.370256 6.325754 6.280857 6.235559
#>  [89] 6.189859 6.143756 6.097252 6.050351 6.003060 5.955386 5.907340 5.858934
#>  [97] 5.810183 5.761102 5.711834 5.662336 5.612542 5.562473 5.512147 5.461586
#> [105] 5.410814 5.359853 5.308729 5.257466 5.206091 5.154630 5.103110 5.051557
#> [113] 5.000000 4.950923 4.901895 4.852910 4.803961 4.755042 4.706146 4.657266
#> [121] 4.608395 4.559525 4.510651 4.461763 4.412856 4.363921 4.314952 4.265940
#> [129] 4.216878 4.167759 4.118576 4.069320 4.019985 3.970563 3.921047 3.871431
#> [137] 3.821706 3.771867 3.719807 3.667342 3.614657 3.561756 3.508648 3.455337
#> [145] 3.401833 3.348143 3.294275 3.240238 3.186040 3.131691 3.077201 3.022579
#> [153] 2.967834 2.912978 2.858021 2.802973 2.747844 2.692646 2.637388 2.582083
#> [161] 2.526740 2.471371 2.415986 2.362867 2.310059 2.257115 2.204034 2.150811
#> [169] 2.097445 2.043935 1.990278 1.936473 1.882520 1.828418 1.774168 1.719769
#> [177] 1.665222 1.610530 1.555692 1.500712 1.445592 1.390334 1.334942 1.279419
#> [185] 1.223770 1.167999 1.112110 1.056108 1.000000 1.060508 1.121076 1.181780
#> [193] 1.242688 1.303866 1.365694 1.428060 1.490824 1.553941 1.617352 1.680682
#> [201] 1.743924 1.807508 1.871411 1.935594 2.000000
#>