Compute a rough multipath outline

Compute a rough multipath outline

Usage

rough_polypath(
  x,
  y = NULL,
  id = NULL,
  rule = c("winding", "evenodd"),
  hand = NULL
)

Arguments

x, y

Coordinates as for graphics::polypath().

id

Optional path ids. Consecutive points with the same id belong to one closed ring.

rule

Fill rule, "winding" or "evenodd".

hand

Hand-drawn geometry settings created with hand().

Value

A list with x, y, id, and rule components describing roughened closed rings.

Examples

rough_polypath(c(2, 4, 4, 2, 2.5, 3.5, 3.5, 2.5),
               c(2, 2, 4, 4, 2.5, 2.5, 3.5, 3.5),
               id = c(rep(1, 4), rep(2, 4)),
               rule = "evenodd")
#> $x
#>   [1] 2.000000 2.086957 2.173913 2.260870 2.347826 2.434783 2.521739 2.608696
#>   [9] 2.695652 2.782609 2.869565 2.956522 3.043478 3.130435 3.217391 3.304348
#>  [17] 3.391304 3.478261 3.565217 3.652174 3.739130 3.826087 3.913043 4.000000
#>  [25] 4.004525 4.009361 4.014406 4.019549 4.024668 4.029638 4.034327 4.036923
#>  [33] 4.036206 4.034527 4.032029 4.028876 4.025243 4.021316 4.017280 4.016538
#>  [41] 4.015997 4.014853 4.013084 4.010688 4.007682 4.004101 4.000000 3.913043
#>  [49] 3.826087 3.739130 3.652174 3.565217 3.478261 3.391304 3.304348 3.217391
#>  [57] 3.130435 3.043478 2.956522 2.869565 2.782609 2.695652 2.608696 2.521739
#>  [65] 2.434783 2.347826 2.260870 2.173913 2.086957 2.000000 1.995406 1.990625
#>  [73] 1.985752 1.980891 1.976150 1.971640 1.967469 1.964825 1.964513 1.965039
#>  [81] 1.966325 1.968275 1.970779 1.973719 1.976971 1.978486 1.980088 1.982229
#>  [89] 1.984895 1.988054 1.991664 1.995669 2.000000 2.500000 2.590909 2.681818
#>  [97] 2.772727 2.863636 2.954545 3.045455 3.136364 3.227273 3.318182 3.409091
#> [105] 3.500000 3.502158 3.504954 3.508061 3.510329 3.510554 3.509868 3.508438
#> [113] 3.506435 3.504179 3.501956 3.500000 3.409091 3.318182 3.227273 3.136364
#> [121] 3.045455 2.954545 2.863636 2.772727 2.681818 2.590909 2.500000 2.495904
#> [129] 2.492387 2.489703 2.487407 2.484943 2.483588 2.483672 2.486660 2.490936
#> [137] 2.495526 2.500000
#> 
#> $y
#>   [1] 2.000000 2.002609 2.005036 2.007240 2.009186 2.010847 2.012201 2.013237
#>   [9] 2.014103 2.014904 2.015420 2.015646 2.015579 2.015223 2.014587 2.013685
#>  [17] 2.012947 2.011979 2.010687 2.009081 2.007179 2.005008 2.002602 2.000000
#>  [25] 2.086957 2.173913 2.260870 2.347826 2.434783 2.521739 2.608696 2.695652
#>  [33] 2.782609 2.869565 2.956522 3.043478 3.130435 3.217391 3.304348 3.391304
#>  [41] 3.478261 3.565217 3.652174 3.739130 3.826087 3.913043 4.000000 3.999679
#>  [49] 3.998462 3.996397 3.993564 3.990073 3.986060 3.981686 3.979479 3.981156
#>  [57] 3.983468 3.986260 3.989367 3.992613 3.995825 3.998830 3.999531 3.999743
#>  [65] 3.999919 4.000049 4.000126 4.000146 4.000104 4.000000 3.913043 3.826087
#>  [73] 3.739130 3.652174 3.565217 3.478261 3.391304 3.304348 3.217391 3.130435
#>  [81] 3.043478 2.956522 2.869565 2.782609 2.695652 2.608696 2.521739 2.434783
#>  [89] 2.347826 2.260870 2.173913 2.086957 2.000000 2.500000 2.510439 2.520253
#>  [97] 2.528620 2.535139 2.539577 2.540919 2.538836 2.531575 2.521662 2.510806
#> [105] 2.500000 2.590909 2.681818 2.772727 2.863636 2.954545 3.045455 3.136364
#> [113] 3.227273 3.318182 3.409091 3.500000 3.492732 3.486464 3.481653 3.477961
#> [121] 3.474861 3.473705 3.474772 3.479031 3.485213 3.492406 3.500000 3.409091
#> [129] 3.318182 3.227273 3.136364 3.045455 2.954545 2.863636 2.772727 2.681818
#> [137] 2.590909 2.500000
#> 
#> $id
#>   [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [75] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#> [112] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#> 
#> $rule
#> [1] "evenodd"
#>