Compute a rough rectangle outline

Compute a rough rectangle outline

Usage

rough_rect(x0, y0, x1, y1, hand = NULL)

Arguments

x0, y0

Rectangle corner.

x1, y1

Opposite rectangle corner.

hand

Hand-drawn geometry settings created with hand().

Value

A list with x and y components containing a roughened closed outline suitable for plotting with graphics::lines().

Examples

rough_rect(2, 2, 5, 6)
#> $x
#>   [1] 2.000000 2.085714 2.171429 2.257143 2.342857 2.428571 2.514286 2.600000
#>   [9] 2.685714 2.771429 2.857143 2.942857 3.028571 3.114286 3.200000 3.285714
#>  [17] 3.371429 3.457143 3.542857 3.628571 3.714286 3.800000 3.885714 3.971429
#>  [25] 4.057143 4.142857 4.228571 4.314286 4.400000 4.485714 4.571429 4.657143
#>  [33] 4.742857 4.828571 4.914286 5.000000 5.002558 5.004992 5.007290 5.009445
#>  [41] 5.011447 5.013291 5.014971 5.016482 5.017822 5.018988 5.019981 5.020799
#>  [49] 5.021446 5.021924 5.022236 5.024965 5.030491 5.036170 5.041935 5.047717
#>  [57] 5.053444 5.059048 5.064456 5.069599 5.074409 5.078818 5.082760 5.086173
#>  [65] 5.088997 5.091177 5.092661 5.085706 5.078504 5.071282 5.064101 5.057021
#>  [73] 5.050100 5.043391 5.036948 5.030819 5.025049 5.019681 5.014750 5.010290
#>  [81] 5.006330 5.002894 5.000000 4.914286 4.828571 4.742857 4.657143 4.571429
#>  [89] 4.485714 4.400000 4.314286 4.228571 4.142857 4.057143 3.971429 3.885714
#>  [97] 3.800000 3.714286 3.628571 3.542857 3.457143 3.371429 3.285714 3.200000
#> [105] 3.114286 3.028571 2.942857 2.857143 2.771429 2.685714 2.600000 2.514286
#> [113] 2.428571 2.342857 2.257143 2.171429 2.085714 2.000000 2.004896 2.009990
#> [121] 2.015258 2.020673 2.026208 2.031835 2.037521 2.043237 2.048949 2.054624
#> [129] 2.060227 2.065725 2.071083 2.076265 2.081237 2.083476 2.082587 2.081096
#> [137] 2.079044 2.076475 2.073436 2.069977 2.066150 2.062009 2.057609 2.053008
#> [145] 2.048262 2.043429 2.038566 2.033729 2.028975 2.029780 2.030394 2.030679
#> [153] 2.030619 2.030197 2.029401 2.028219 2.026644 2.024673 2.022302 2.019533
#> [161] 2.016372 2.012826 2.008905 2.004624 2.000000
#> 
#> $y
#>   [1] 2.000000 2.004694 2.009675 2.014897 2.020313 2.025870 2.031511 2.037178
#>   [9] 2.042809 2.048340 2.053706 2.058843 2.062359 2.063706 2.064433 2.064556
#>  [17] 2.064098 2.063087 2.061556 2.059542 2.057085 2.054230 2.051022 2.047513
#>  [25] 2.044697 2.041659 2.038316 2.034698 2.030838 2.026769 2.022527 2.018147
#>  [33] 2.013667 2.009125 2.004556 2.000000 2.085106 2.170213 2.255319 2.340426
#>  [41] 2.425532 2.510638 2.595745 2.680851 2.765957 2.851064 2.936170 3.021277
#>  [49] 3.106383 3.191489 3.276596 3.361702 3.446809 3.531915 3.617021 3.702128
#>  [57] 3.787234 3.872340 3.957447 4.042553 4.127660 4.212766 4.297872 4.382979
#>  [65] 4.468085 4.553191 4.638298 4.723404 4.808511 4.893617 4.978723 5.063830
#>  [73] 5.148936 5.234043 5.319149 5.404255 5.489362 5.574468 5.659574 5.744681
#>  [81] 5.829787 5.914894 6.000000 5.998984 5.998424 5.998320 5.998663 5.999439
#>  [89] 6.000624 6.002191 6.004104 6.006323 6.008802 6.011491 6.011527 6.007827
#>  [97] 6.003742 5.999373 5.994826 5.990211 5.985638 5.981217 5.977058 5.973266
#> [105] 5.969943 5.967185 5.970025 5.973488 5.976940 5.980329 5.983606 5.986724
#> [113] 5.989640 5.992315 5.994715 5.996811 5.998579 6.000000 5.914894 5.829787
#> [121] 5.744681 5.659574 5.574468 5.489362 5.404255 5.319149 5.234043 5.148936
#> [129] 5.063830 4.978723 4.893617 4.808511 4.723404 4.638298 4.553191 4.468085
#> [137] 4.382979 4.297872 4.212766 4.127660 4.042553 3.957447 3.872340 3.787234
#> [145] 3.702128 3.617021 3.531915 3.446809 3.361702 3.276596 3.191489 3.106383
#> [153] 3.021277 2.936170 2.851064 2.765957 2.680851 2.595745 2.510638 2.425532
#> [161] 2.340426 2.255319 2.170213 2.085106 2.000000
#>