Compute a rough polygon outline

Compute a rough polygon outline

Usage

rough_polygon(x, y = NULL, hand = NULL)

Arguments

x, y

Polygon coordinates.

hand

Hand-drawn geometry settings created with hand().

Value

A list with x and y components containing a roughened closed outline suitable for plotting with graphics::lines().

Examples

rough_polygon(c(2, 5, 8, 3), c(2, 7, 5, 1))
#> $x
#>   [1] 2.000000 2.048167 2.096232 2.144186 2.192021 2.239729 2.287300 2.334728
#>   [9] 2.382006 2.429128 2.476086 2.522877 2.569495 2.615935 2.662194 2.708268
#>  [17] 2.754155 2.799852 2.845357 2.890670 2.935789 2.980715 3.025447 3.070563
#>  [25] 3.117494 3.164351 3.211118 3.257784 3.304332 3.350751 3.397026 3.443143
#>  [33] 3.489090 3.534854 3.580423 3.625783 3.670924 3.715834 3.760503 3.804919
#>  [41] 3.849075 3.892960 3.936566 3.979886 4.022913 4.065640 4.106975 4.146817
#>  [49] 4.186486 4.225993 4.265350 4.304568 4.343657 4.382630 4.421499 4.460274
#>  [57] 4.498969 4.537595 4.576163 4.614687 4.653177 4.691646 4.730106 4.768567
#>  [65] 4.807041 4.845541 4.884075 4.922657 4.961295 5.000000 5.064808 5.129342
#>  [73] 5.193632 5.257712 5.321617 5.385387 5.449060 5.512681 5.576293 5.639942
#>  [81] 5.703675 5.767538 5.831581 5.895851 5.963390 6.032186 6.101550 6.171463
#>  [89] 6.241904 6.312847 6.384264 6.456123 6.528391 6.601033 6.674010 6.747285
#>  [97] 6.820817 6.894565 6.967364 7.039748 7.112417 7.185359 7.258554 7.331987
#> [105] 7.405637 7.479484 7.553505 7.627680 7.701984 7.776392 7.850881 7.925426
#> [113] 8.000000 7.933212 7.866437 7.799677 7.732932 7.666206 7.599498 7.532811
#> [121] 7.466147 7.399505 7.332889 7.266298 7.199735 7.133200 7.066694 7.000218
#> [129] 6.933774 6.867361 6.800982 6.734636 6.668324 6.602047 6.535805 6.469598
#> [137] 6.403428 6.337294 6.270857 6.204411 6.137988 6.071590 6.005219 5.938875
#> [145] 5.872562 5.806282 5.740035 5.673823 5.607649 5.541514 5.475420 5.409367
#> [153] 5.343358 5.277394 5.211475 5.145604 5.079782 5.014008 4.948285 4.882612
#> [161] 4.816992 4.751424 4.685908 4.620756 4.555699 4.490674 4.425682 4.360719
#> [169] 4.295784 4.230875 4.165990 4.101128 4.036286 3.971463 3.906656 3.841863
#> [177] 3.777083 3.712314 3.647554 3.582800 3.518051 3.453305 3.388559 3.323813
#> [185] 3.259064 3.194310 3.129549 3.064779 3.000000 2.936637 2.873542 2.810722
#> [193] 2.748170 2.685856 2.623084 2.560102 2.497184 2.434374 2.371712 2.309283
#> [201] 2.247100 2.185092 2.123252 2.061564 2.000000
#> 
#> $y
#>   [1] 2.000000 2.069651 2.139362 2.209140 2.278990 2.348916 2.418924 2.489018
#>   [9] 2.559202 2.629480 2.699855 2.770332 2.840912 2.911598 2.982394 3.053300
#>  [17] 3.124319 3.195451 3.266699 3.338062 3.409541 3.481136 3.552848 3.624329
#>  [25] 3.694721 3.765158 3.835648 3.906199 3.976821 4.047520 4.118306 4.189187
#>  [33] 4.260169 4.331261 4.402471 4.473806 4.545272 4.616876 4.688626 4.760527
#>  [41] 4.832584 4.904804 4.977191 5.049749 5.122484 5.195398 5.269149 5.343794
#>  [49] 5.418543 5.493389 5.568326 5.643346 5.718443 5.793610 5.868840 5.944125
#>  [57] 6.019459 6.094835 6.170244 6.245681 6.321137 6.396606 6.472082 6.547556
#>  [65] 6.623022 6.698473 6.773903 6.849305 6.924672 7.000000 6.946049 6.891687
#>  [73] 6.836960 6.781917 6.726612 6.671103 6.615451 6.559719 6.503974 6.448285
#>  [81] 6.392721 6.337354 6.282256 6.227497 6.177643 6.129674 6.082557 6.036265
#>  [89] 5.990764 5.946015 5.901977 5.858603 5.815842 5.773642 5.731946 5.690695
#>  [97] 5.649830 5.609289 5.567326 5.524738 5.482579 5.440829 5.399460 5.358446
#> [105] 5.317758 5.277365 5.237235 5.197334 5.157626 5.118077 5.078648 5.039301
#> [113] 5.000000 4.948616 4.897217 4.845799 4.794361 4.742901 4.691417 4.639907
#> [121] 4.588369 4.536803 4.485205 4.433575 4.381911 4.330211 4.278475 4.226701
#> [129] 4.174888 4.123035 4.071141 4.019205 3.967227 3.915205 3.863139 3.811028
#> [137] 3.758873 3.706672 3.654850 3.603039 3.551199 3.499328 3.447424 3.395485
#> [145] 3.343507 3.291490 3.239430 3.187326 3.135175 3.082975 3.030725 2.978422
#> [153] 2.926065 2.873652 2.821182 2.768652 2.716063 2.663411 2.610697 2.557919
#> [161] 2.505076 2.452168 2.399194 2.345766 2.292219 2.238631 2.185003 2.131338
#> [169] 2.077638 2.023906 1.970144 1.916353 1.862537 1.808698 1.754838 1.700961
#> [177] 1.647067 1.593160 1.539242 1.485316 1.431384 1.377448 1.323511 1.269576
#> [185] 1.215644 1.161718 1.107801 1.053894 1.000000 1.061637 1.123542 1.185722
#> [193] 1.248170 1.310856 1.373084 1.435102 1.497184 1.559374 1.621712 1.684283
#> [201] 1.747100 1.810092 1.873252 1.936564 2.000000
#>