Compute a rough polygon outline

Compute a rough polygon outline

Usage

rough_polygon(x, y = NULL, hand = NULL)

Arguments

x, y

Polygon coordinates.

hand

Hand-drawn geometry settings created with hand().

Value

A list with x and y components containing a roughened closed outline suitable for plotting with graphics::lines().

Examples

rough_polygon(c(2, 5, 8, 3), c(2, 7, 5, 1))
#> $x
#>   [1] 1.949077 1.989938 2.030654 2.071234 2.111688 2.152024 2.192254 2.232386
#>   [9] 2.272434 2.312407 2.352319 2.392182 2.432008 2.471811 2.511604 2.551401
#>  [17] 2.591217 2.631064 2.670959 2.710915 2.750947 2.791070 2.831298 2.872609
#>  [25] 2.917406 2.962545 3.008010 3.053790 3.099868 3.146228 3.192856 3.239733
#>  [33] 3.286841 3.334163 3.381679 3.429370 3.477217 3.525199 3.573297 3.621489
#>  [41] 3.669756 3.718077 3.766431 3.814798 3.863158 3.911491 3.957938 4.002320
#>  [49] 4.046855 4.091548 4.136404 4.181426 4.226619 4.271984 4.317526 4.363244
#>  [57] 4.409141 4.455217 4.501472 4.547905 4.594517 4.641303 4.688263 4.735393
#>  [65] 4.782689 4.830148 4.877764 4.925533 4.973447 5.021500 5.067495 5.136376
#>  [73] 5.205270 5.274190 5.343151 5.412163 5.481242 5.550399 5.619647 5.688997
#>  [81] 5.758461 5.828050 5.897774 5.967643 6.036951 6.106198 6.175566 6.245076
#>  [89] 6.314750 6.384607 6.454668 6.524949 6.595470 6.666246 6.737292 6.808620
#>  [97] 6.880241 6.952166 7.025581 7.099889 7.174311 7.248822 7.323396 7.398007
#> [105] 7.472630 7.547240 7.621815 7.696331 7.770767 7.845102 7.919317 7.993395
#> [113] 8.067321 7.830029 7.762847 7.695736 7.628701 7.561746 7.494874 7.428090
#> [121] 7.361396 7.294795 7.228291 7.161886 7.095581 7.029380 6.963284 6.897293
#> [129] 6.831411 6.765637 6.699971 6.634415 6.568969 6.503632 6.438404 6.373284
#> [137] 6.308272 6.243366 6.176898 6.110314 6.043763 5.977252 5.910788 5.844381
#> [145] 5.778037 5.711763 5.645567 5.579458 5.513440 5.447523 5.381713 5.316016
#> [153] 5.250440 5.184990 5.119674 5.054496 4.989462 4.924578 4.859848 4.795279
#> [161] 4.730873 4.666635 4.602568 4.540008 4.477804 4.415693 4.353669 4.291725
#> [169] 4.229857 4.168058 4.106323 4.044646 3.983021 3.921442 3.859902 3.798397
#> [177] 3.736919 3.675463 3.614023 3.552593 3.491168 3.429740 3.368306 3.306858
#> [185] 3.245392 3.183902 3.122382 3.060828 2.999235 2.917462 2.854822 2.792294
#> [193] 2.729834 2.667387 2.602934 2.537324 2.471517 2.405799 2.340457 2.277408
#> [201] 2.216517 2.155229 2.093371 2.030813 1.967482
#> 
#> $y
#>   [1] 1.895944 1.973120 2.050383 2.127727 2.205147 2.282638 2.360193 2.437806
#>   [9] 2.515470 2.593178 2.670923 2.748698 2.826495 2.904306 2.982122 3.059936
#>  [17] 3.137740 3.215523 3.293279 3.370998 3.448671 3.526290 3.603846 3.680752
#>  [25] 3.755566 3.830175 3.904588 3.978813 4.052859 4.126735 4.200451 4.274017
#>  [33] 4.347445 4.420744 4.493927 4.567004 4.639989 4.712892 4.785726 4.858503
#>  [41] 4.931235 5.003935 5.076615 5.149287 5.221964 5.294656 5.368481 5.443544
#>  [49] 5.518516 5.593392 5.668171 5.742850 5.817427 5.891900 5.966268 6.040529
#>  [57] 6.114684 6.188731 6.262670 6.336502 6.410228 6.483849 6.557365 6.630780
#>  [65] 6.704094 6.777311 6.850434 6.923466 6.996410 7.069270 6.941807 6.892843
#>  [73] 6.843898 6.794993 6.746147 6.697381 6.648713 6.600163 6.551748 6.503488
#>  [81] 6.455398 6.407496 6.359796 6.312313 6.263990 6.215575 6.167341 6.119321
#>  [89] 6.071545 6.024045 5.976850 5.929987 5.883483 5.837361 5.791643 5.746349
#>  [97] 5.701496 5.657098 5.614934 5.574109 5.533457 5.492938 5.452513 5.412143
#> [105] 5.371792 5.331422 5.290999 5.250487 5.209854 5.169071 5.128108 5.086940
#> [113] 5.045543 4.905280 4.856796 4.808223 4.759555 4.710788 4.661916 4.612935
#> [121] 4.563841 4.514631 4.465299 4.415845 4.366264 4.316554 4.266713 4.216739
#> [129] 4.166631 4.116387 4.066007 4.015491 3.964837 3.914047 3.863121 3.812059
#> [137] 3.760863 3.709534 3.660157 3.610926 3.561654 3.512331 3.462949 3.413497
#> [145] 3.363966 3.314346 3.264629 3.214805 3.164865 3.114800 3.064601 3.014261
#> [153] 2.963770 2.913120 2.862305 2.811316 2.760147 2.708790 2.657240 2.605491
#> [161] 2.553537 2.501373 2.448995 2.394734 2.340027 2.285204 2.230274 2.175242
#> [169] 2.120115 2.064902 2.009609 1.954244 1.898814 1.843327 1.787790 1.732210
#> [177] 1.676596 1.620955 1.565293 1.509619 1.453939 1.398262 1.342594 1.286942
#> [185] 1.231313 1.175715 1.120152 1.064633 1.009164 1.054500 1.118369 1.182350
#> [193] 1.246399 1.310461 1.372517 1.433415 1.494117 1.554909 1.616075 1.679535
#> [201] 1.745152 1.810374 1.875024 1.938975 2.002154
#>