Compute a rough multipath outline

Compute a rough multipath outline

Usage

rough_polypath(
  x,
  y = NULL,
  id = NULL,
  rule = c("winding", "evenodd"),
  hand = NULL
)

Arguments

x, y

Coordinates as for graphics::polypath().

id

Optional path ids. Consecutive points with the same id belong to one closed ring.

rule

Fill rule, "winding" or "evenodd".

hand

Hand-drawn geometry settings created with hand().

Value

A list with x, y, id, and rule components describing roughened closed rings.

Examples

rough_polypath(c(2, 4, 4, 2, 2.5, 3.5, 3.5, 2.5),
               c(2, 2, 4, 4, 2.5, 2.5, 3.5, 3.5),
               id = c(rep(1, 4), rep(2, 4)),
               rule = "evenodd")
#> $x
#>   [1] 1.982685 2.070095 2.157506 2.244916 2.332326 2.419737 2.507147 2.594557
#>   [9] 2.681968 2.769378 2.856789 2.944199 3.031609 3.119020 3.206430 3.293840
#>  [17] 3.381251 3.468661 3.556072 3.643482 3.730892 3.818303 3.905713 3.993123
#>  [25] 3.975046 3.972102 3.969234 3.966521 3.964038 3.961859 3.960051 3.959285
#>  [33] 3.960012 3.961386 3.963352 3.965844 3.968784 3.972087 3.975662 3.977770
#>  [41] 3.979987 3.982717 3.985957 3.989691 3.993887 3.998501 4.003476 3.931998
#>  [49] 3.844488 3.756978 3.669467 3.581957 3.494447 3.406936 3.319426 3.231916
#>  [57] 3.144405 3.056895 2.969385 2.881875 2.794364 2.706854 2.619344 2.531833
#>  [65] 2.444323 2.356813 2.269302 2.181792 2.094282 2.006772 2.003239 2.008519
#>  [73] 2.013768 2.018882 2.023759 2.028297 2.032397 2.035535 2.037347 2.038393
#>  [81] 2.038677 2.038216 2.037044 2.035205 2.032757 2.030132 2.027035 2.023441
#>  [89] 2.019419 2.015047 2.010409 2.005595 2.000695 2.509308 2.598749 2.688189
#>  [97] 2.777630 2.867070 2.956510 3.045951 3.135391 3.224831 3.314272 3.403712
#> [105] 3.493153 3.514217 3.514949 3.515348 3.515147 3.514114 3.512126 3.509115
#> [113] 3.504595 3.499328 3.493833 3.488410 3.417649 3.326846 3.236042 3.145239
#> [121] 3.054436 2.963632 2.872829 2.782026 2.691222 2.600419 2.509616 2.507876
#> [129] 2.514565 2.519445 2.522669 2.524609 2.524334 2.521768 2.517276 2.510820
#> [137] 2.502686 2.493435
#> 
#> $y
#>   [1] 2.021258 2.019176 2.017898 2.017422 2.017718 2.018727 2.020366 2.022527
#>   [9] 2.022947 2.020096 2.016956 2.013688 2.010460 2.007441 2.004794 2.002672
#>  [17] 2.003343 2.004766 2.006397 2.008199 2.010129 2.012145 2.014204 2.016261
#>  [25] 2.053457 2.141180 2.228903 2.316625 2.404348 2.492071 2.579794 2.667517
#>  [33] 2.755240 2.842962 2.930685 3.018408 3.106131 3.193854 3.281577 3.369300
#>  [41] 3.457022 3.544745 3.632468 3.720191 3.807914 3.895637 3.983360 3.981628
#>  [49] 3.986428 3.991272 3.996071 4.000735 4.005174 4.009299 4.012558 4.014566
#>  [57] 4.015935 4.016669 4.016786 4.016316 4.015300 4.013788 4.012059 4.009954
#>  [65] 4.007478 4.004695 4.001672 3.998483 3.995199 3.991898 3.906976 3.820611
#>  [73] 3.734246 3.647880 3.561515 3.475150 3.388785 3.302420 3.216054 3.129689
#>  [81] 3.043324 2.956959 2.870594 2.784228 2.697863 2.611498 2.525133 2.438768
#>  [89] 2.352403 2.266037 2.179672 2.093307 2.006942 2.503374 2.503160 2.503742
#>  [97] 2.504917 2.506355 2.507659 2.508460 2.508398 2.504703 2.500495 2.497424
#> [105] 2.495879 2.592836 2.685586 2.778337 2.871088 2.963838 3.056589 3.149339
#> [113] 3.242090 3.334841 3.427591 3.520342 3.485325 3.482434 3.481439 3.480734
#> [121] 3.478442 3.477757 3.479402 3.486132 3.495218 3.505054 3.514961 3.416156
#> [129] 3.326175 3.236194 3.146213 3.056232 2.966251 2.876270 2.786289 2.696308
#> [137] 2.606327 2.516346
#> 
#> $id
#>   [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [75] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#> [112] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#> 
#> $rule
#> [1] "evenodd"
#>