Compute a rough multipath outline

Compute a rough multipath outline

Usage

rough_polypath(
  x,
  y = NULL,
  id = NULL,
  rule = c("winding", "evenodd"),
  hand = NULL
)

Arguments

x, y

Coordinates as for graphics::polypath().

id

Optional path ids. Consecutive points with the same id belong to one closed ring.

rule

Fill rule, "winding" or "evenodd".

hand

Hand-drawn geometry settings created with hand().

Value

A list with x, y, id, and rule components describing roughened closed rings.

Examples

rough_polypath(c(2, 4, 4, 2, 2.5, 3.5, 3.5, 2.5),
               c(2, 2, 4, 4, 2.5, 2.5, 3.5, 3.5),
               id = c(rep(1, 4), rep(2, 4)),
               rule = "evenodd")
#> $x
#>   [1] 2.000000 2.086957 2.173913 2.260870 2.347826 2.434783 2.521739 2.608696
#>   [9] 2.695652 2.782609 2.869565 2.956522 3.043478 3.130435 3.217391 3.304348
#>  [17] 3.391304 3.478261 3.565217 3.652174 3.739130 3.826087 3.913043 4.000000
#>  [25] 3.996244 3.992690 3.989399 3.986429 3.983825 3.981625 3.979859 3.978037
#>  [33] 3.975840 3.974010 3.972615 3.971712 3.971354 3.971579 3.972416 3.974756
#>  [41] 3.977713 3.981009 3.984574 3.988333 3.992208 3.996122 4.000000 3.913043
#>  [49] 3.826087 3.739130 3.652174 3.565217 3.478261 3.391304 3.304348 3.217391
#>  [57] 3.130435 3.043478 2.956522 2.869565 2.782609 2.695652 2.608696 2.521739
#>  [65] 2.434783 2.347826 2.260870 2.173913 2.086957 2.000000 2.001424 2.002699
#>  [73] 2.003806 2.004729 2.005459 2.005991 2.006328 2.007103 2.008735 2.010320
#>  [81] 2.011784 2.013051 2.014052 2.014720 2.014999 2.013456 2.011456 2.009392
#>  [89] 2.007319 2.005296 2.003372 2.001594 2.000000 2.500000 2.590909 2.681818
#>  [97] 2.772727 2.863636 2.954545 3.045455 3.136364 3.227273 3.318182 3.409091
#> [105] 3.500000 3.495581 3.490834 3.486227 3.482545 3.480569 3.480132 3.481340
#> [113] 3.485339 3.490411 3.495471 3.500000 3.409091 3.318182 3.227273 3.136364
#> [121] 3.045455 2.954545 2.863636 2.772727 2.681818 2.590909 2.500000 2.508862
#> [129] 2.516075 2.521169 2.524435 2.526417 2.526246 2.523962 2.520903 2.515948
#> [137] 2.508829 2.500000
#> 
#> $y
#>   [1] 2.000000 2.003371 2.007023 2.010877 2.014846 2.018834 2.022741 2.026460
#>   [9] 2.027866 2.025544 2.022356 2.018514 2.014248 2.009797 2.005399 2.001291
#>  [17] 2.002513 2.004379 2.005621 2.006143 2.005872 2.004763 2.002801 2.000000
#>  [25] 2.086957 2.173913 2.260870 2.347826 2.434783 2.521739 2.608696 2.695652
#>  [33] 2.782609 2.869565 2.956522 3.043478 3.130435 3.217391 3.304348 3.391304
#>  [41] 3.478261 3.565217 3.652174 3.739130 3.826087 3.913043 4.000000 4.007484
#>  [49] 4.015165 4.022891 4.030501 4.037833 4.044724 4.051013 4.056292 4.060253
#>  [57] 4.063142 4.064884 4.065428 4.064743 4.062821 4.059680 4.053829 4.046883
#>  [65] 4.039387 4.031524 4.023479 4.015431 4.007551 4.000000 3.913043 3.826087
#>  [73] 3.739130 3.652174 3.565217 3.478261 3.391304 3.304348 3.217391 3.130435
#>  [81] 3.043478 2.956522 2.869565 2.782609 2.695652 2.608696 2.521739 2.434783
#>  [89] 2.347826 2.260870 2.173913 2.086957 2.000000 2.500000 2.499112 2.497425
#>  [97] 2.495183 2.494528 2.497719 2.501391 2.504653 2.503455 2.501370 2.500139
#> [105] 2.500000 2.590909 2.681818 2.772727 2.863636 2.954545 3.045455 3.136364
#> [113] 3.227273 3.318182 3.409091 3.500000 3.502597 3.506122 3.510151 3.512887
#> [121] 3.512437 3.510850 3.508513 3.506484 3.504373 3.502141 3.500000 3.409091
#> [129] 3.318182 3.227273 3.136364 3.045455 2.954545 2.863636 2.772727 2.681818
#> [137] 2.590909 2.500000
#> 
#> $id
#>   [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [75] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#> [112] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#> 
#> $rule
#> [1] "evenodd"
#>