Compute a rough rectangle outline

Compute a rough rectangle outline

Usage

rough_rect(x0, y0, x1, y1, hand = NULL)

Arguments

x0, y0

Rectangle corner.

x1, y1

Opposite rectangle corner.

hand

Hand-drawn geometry settings created with hand().

Value

A list with x and y components containing a roughened closed outline suitable for plotting with graphics::lines().

Examples

rough_rect(2, 2, 5, 6)
#> $x
#>   [1] 2.019097 2.104706 2.190316 2.275925 2.361535 2.447144 2.532754 2.618363
#>   [9] 2.703973 2.789582 2.875192 2.960802 3.046411 3.132021 3.217630 3.303240
#>  [17] 3.388849 3.474459 3.560068 3.645678 3.731287 3.816897 3.902506 3.988116
#>  [25] 4.073725 4.159335 4.244944 4.330554 4.416163 4.501773 4.587382 4.672992
#>  [33] 4.758601 4.844211 4.929820 5.015430 5.115231 5.113043 5.110990 5.109067
#>  [41] 5.107264 5.105575 5.103989 5.102494 5.101080 5.099733 5.098440 5.097187
#>  [49] 5.095958 5.094738 5.093510 5.092591 5.092010 5.091412 5.090768 5.090054
#>  [57] 5.089241 5.088304 5.087217 5.085954 5.084490 5.082801 5.080865 5.078660
#>  [65] 5.076166 5.073365 5.070240 5.062897 5.055477 5.048100 5.040805 5.033632
#>  [73] 5.026614 5.019786 5.013180 5.006825 5.000747 4.994971 4.989517 4.984404
#>  [81] 4.979646 4.975255 4.971239 4.888388 4.803035 4.717682 4.632329 4.546976
#>  [89] 4.461623 4.376270 4.290917 4.205564 4.120211 4.034858 3.949505 3.864152
#>  [97] 3.778799 3.693446 3.608092 3.522739 3.437386 3.352033 3.266680 3.181327
#> [105] 3.095974 3.010621 2.925268 2.839915 2.754562 2.669209 2.583856 2.498503
#> [113] 2.413150 2.327797 2.242444 2.157091 2.071738 1.986385 2.000905 1.996915
#> [121] 1.992656 1.988145 1.983398 1.978435 1.973277 1.967951 1.962480 1.956893
#> [129] 1.951219 1.945489 1.939735 1.933990 1.928288 1.925006 1.924519 1.924453
#> [137] 1.924773 1.925444 1.926426 1.927680 1.929164 1.930835 1.932649 1.934561
#> [145] 1.936528 1.938504 1.940445 1.942307 1.944049 1.942044 1.940090 1.938272
#> [153] 1.936599 1.935078 1.933715 1.932516 1.931483 1.930621 1.929930 1.929410
#> [161] 1.929060 1.928876 1.928855 1.928991 1.929278
#> 
#> $y
#>   [1] 1.947459 1.955079 1.962719 1.970334 1.977879 1.985310 1.992585 1.999659
#>   [9] 2.006491 2.013040 2.019269 2.025141 2.030706 2.035964 2.040785 2.045142
#>  [17] 2.049008 2.052362 2.055186 2.057467 2.059195 2.060367 2.060982 2.061046
#>  [25] 2.059488 2.057427 2.055035 2.052361 2.049453 2.046363 2.043140 2.039837
#>  [33] 2.036504 2.033189 2.029942 2.026808 2.113266 2.196892 2.280519 2.364146
#>  [41] 2.447773 2.531400 2.615027 2.698654 2.782281 2.865908 2.949534 3.033161
#>  [49] 3.116788 3.200415 3.284042 3.367669 3.451296 3.534923 3.618549 3.702176
#>  [57] 3.785803 3.869430 3.953057 4.036684 4.120311 4.203938 4.287565 4.371191
#>  [65] 4.454818 4.538445 4.622072 4.705699 4.789326 4.872953 4.956580 5.040207
#>  [73] 5.123833 5.207460 5.291087 5.374714 5.458341 5.541968 5.625595 5.709222
#>  [81] 5.792849 5.876475 5.960102 6.023838 6.021838 6.019531 6.016933 6.014064
#>  [89] 6.010948 6.007614 6.004095 6.000428 5.996653 5.992812 5.990839 5.991470
#>  [97] 5.992458 5.993740 5.995250 5.996919 5.998675 6.000448 6.002166 6.003757
#> [105] 6.005153 6.006289 6.004009 6.001417 5.998904 5.996492 5.994204 5.992060
#> [113] 5.990078 5.988273 5.986657 5.985238 5.984023 5.983014 5.960952 5.875261
#> [121] 5.789570 5.703880 5.618189 5.532499 5.446808 5.361118 5.275427 5.189737
#> [129] 5.104046 5.018356 4.932665 4.846975 4.761284 4.675593 4.589903 4.504212
#> [137] 4.418522 4.332831 4.247141 4.161450 4.075760 3.990069 3.904379 3.818688
#> [145] 3.732998 3.647307 3.561616 3.475926 3.390235 3.304545 3.218854 3.133164
#> [153] 3.047473 2.961783 2.876092 2.790402 2.704711 2.619021 2.533330 2.447639
#> [161] 2.361949 2.276258 2.190568 2.104877 2.019187
#>