Compute a rough rectangle outline

Compute a rough rectangle outline

Usage

rough_rect(x0, y0, x1, y1, hand = NULL)

Arguments

x0, y0

Rectangle corner.

x1, y1

Opposite rectangle corner.

hand

Hand-drawn geometry settings created with hand().

Value

A list with x and y components containing a roughened closed outline suitable for plotting with graphics::lines().

Examples

rough_rect(2, 2, 5, 6)
#> $x
#>   [1] 2.000000 2.085714 2.171429 2.257143 2.342857 2.428571 2.514286 2.600000
#>   [9] 2.685714 2.771429 2.857143 2.942857 3.028571 3.114286 3.200000 3.285714
#>  [17] 3.371429 3.457143 3.542857 3.628571 3.714286 3.800000 3.885714 3.971429
#>  [25] 4.057143 4.142857 4.228571 4.314286 4.400000 4.485714 4.571429 4.657143
#>  [33] 4.742857 4.828571 4.914286 5.000000 4.992028 4.984124 4.976323 4.968658
#>  [41] 4.961164 4.953873 4.946818 4.940029 4.933535 4.927364 4.921543 4.916095
#>  [49] 4.911045 4.906412 4.902217 4.896565 4.889194 4.882081 4.875293 4.868896
#>  [57] 4.862954 4.857530 4.852687 4.848483 4.844974 4.842214 4.840252 4.839134
#>  [65] 4.838900 4.839589 4.841230 4.850506 4.860315 4.870442 4.880816 4.891365
#>  [73] 4.902017 4.912703 4.923355 4.933906 4.944293 4.954456 4.964336 4.973879
#>  [81] 4.983034 4.991755 5.000000 4.914286 4.828571 4.742857 4.657143 4.571429
#>  [89] 4.485714 4.400000 4.314286 4.228571 4.142857 4.057143 3.971429 3.885714
#>  [97] 3.800000 3.714286 3.628571 3.542857 3.457143 3.371429 3.285714 3.200000
#> [105] 3.114286 3.028571 2.942857 2.857143 2.771429 2.685714 2.600000 2.514286
#> [113] 2.428571 2.342857 2.257143 2.171429 2.085714 2.000000 1.998324 1.996486
#> [121] 1.994494 1.992361 1.990097 1.987717 1.985235 1.982667 1.980030 1.977342
#> [129] 1.974622 1.971889 1.969163 1.966465 1.963816 1.961946 1.960982 1.960212
#> [137] 1.959639 1.959260 1.959076 1.959082 1.959277 1.959655 1.960211 1.960940
#> [145] 1.961835 1.962889 1.964094 1.965441 1.966921 1.968930 1.971024 1.973179
#> [153] 1.975378 1.977607 1.979849 1.982091 1.984316 1.986512 1.988664 1.990760
#> [161] 1.992788 1.994736 1.996593 1.998351 2.000000
#> 
#> $y
#>   [1] 2.000000 1.993455 1.987388 1.981842 1.976852 1.972449 1.968651 1.965472
#>   [9] 1.962915 1.960979 1.959651 1.958913 1.957199 1.953914 1.950894 1.948185
#>  [17] 1.945833 1.943880 1.942367 1.941331 1.940803 1.940812 1.941381 1.942530
#>  [25] 1.944777 1.947612 1.950962 1.954810 1.959136 1.963912 1.969108 1.974688
#>  [33] 1.980613 1.986838 1.993317 2.000000 2.085106 2.170213 2.255319 2.340426
#>  [41] 2.425532 2.510638 2.595745 2.680851 2.765957 2.851064 2.936170 3.021277
#>  [49] 3.106383 3.191489 3.276596 3.361702 3.446809 3.531915 3.617021 3.702128
#>  [57] 3.787234 3.872340 3.957447 4.042553 4.127660 4.212766 4.297872 4.382979
#>  [65] 4.468085 4.553191 4.638298 4.723404 4.808511 4.893617 4.978723 5.063830
#>  [73] 5.148936 5.234043 5.319149 5.404255 5.489362 5.574468 5.659574 5.744681
#>  [81] 5.829787 5.914894 6.000000 5.995557 5.990831 5.985865 5.980705 5.975401
#>  [89] 5.970008 5.964583 5.959184 5.953873 5.948713 5.943767 5.939815 5.937182
#>  [97] 5.935056 5.933454 5.932390 5.931870 5.931900 5.932477 5.933598 5.935251
#> [105] 5.937425 5.940100 5.944510 5.949346 5.954399 5.959610 5.964915 5.970257
#> [113] 5.975577 5.980819 5.985929 5.990860 5.995564 6.000000 5.914894 5.829787
#> [121] 5.744681 5.659574 5.574468 5.489362 5.404255 5.319149 5.234043 5.148936
#> [129] 5.063830 4.978723 4.893617 4.808511 4.723404 4.638298 4.553191 4.468085
#> [137] 4.382979 4.297872 4.212766 4.127660 4.042553 3.957447 3.872340 3.787234
#> [145] 3.702128 3.617021 3.531915 3.446809 3.361702 3.276596 3.191489 3.106383
#> [153] 3.021277 2.936170 2.851064 2.765957 2.680851 2.595745 2.510638 2.425532
#> [161] 2.340426 2.255319 2.170213 2.085106 2.000000
#>